webSocket初步构建

+++ src/main/java/com/example/crossafter/goods/bean/Evaluation.java
@@ -3,7 +3,7 @@ package com.example.crossafter.goods.bean;
 public class Evaluation {
     private int gid;
     private double evaluation;
-    private int weight;
+    private double weight;
     private int amount;
 
     public int getGid() {
@@ -22,11 +22,11 @@ public class Evaluation {
         this.evaluation = evaluation;
     }
 
-    public int getWeight() {
+    public double getWeight() {
         return weight;
     }
 
-    public void setWeight(int weight) {
+    public void setWeight(double weight) {
         this.weight = weight;
     }
